Remove unnecessary conversion

+++ python/src/scipp/compat/mantid.py
@@ -64,7 +64,7 @@ def make_detector_info(ws):
             spectrum_[det] = i
             has_spectrum_[det] = True
     detector = sc.Variable([sc.Dim.Detector],
-                           values=det_info.detectorIDs().astype(np.int32))
+                           values=det_info.detectorIDs())
 
     # Remove any information about detectors without data (a spectrum). This
     # mostly just gets in the way and including it in the default converter
